+++ CHANGELOG.md
@@ -4,6 +4,11 @@ All notable changes to this project will be documented in this file.
 
 ## [Unreleased]
 
+### Fixed
+
+- PaaS siren: IPC devices omit channel params; non-IPC devices use `channels` array instead of `channelId`
+- IoT SirenStart `clientLocalTime` format corrected to `yyyyMMdd'T'HHmmss`
+
 ## [1.3.3]
 
 ### Added